+++ lib/shameless/configuration.rb
@@ -2,6 +2,8 @@ module Shameless
   class Configuration
     attr_accessor :partition_urls, :shards_count, :connection_options, :database_extensions
 
+    attr_accessor :legacy_created_at_is_bigint
+
     def shards_per_partition_count
       shards_count / partitions_count
     end