+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add `Configuration#legacy_created_at_is_bigint`
 * Keep a reference to only one model class per table name
 * Make `Store#find_shard` public
 * Make `Store#each_shard` public